// client/src/tools/Samples.model.ts
// Generated by https://quicktype.io

// JSON data datatypes
export interface SamplesData {
    samples: Sample[];
}

export interface Sample {
    id:          string;
    name:        string;
    description: string;
    url:         string;
    images:      Image[];
}

export interface Image {
    filename: string;
}

// component property datatypes
export interface ViewProps {
    samples: Sample[]; 
}